feat: add token refreshing logic in getAccessToken function
The changes in this commit add logic to refresh the access token if it has expired. The function now checks the expiration time of the token and if it is expired, it makes a request to the Spotify API to get a new access token using the client credentials grant type. The new access token is then stored in the database and returned for further use. This ensures that the application always has a valid access token to make API requests to Spotify.